Add capture format and fix framerate

+++ libs/mrwebrtc/src/media/device_video_track_source.h
@@ -26,7 +26,7 @@ struct VideoCaptureDeviceInfo {
 struct VideoCaptureFormatInfo {
   int width;
   int height;
-  int framerate;
+  double framerate;
   uint32_t fourcc;
 };
